etl/etl.py
#!/usr/bin/env python3
"""
Script ETL simple pour le dashboard analytics
"""

import pandas as pd
import psycopg2
from datetime import datetime
import os

def main():
    print(f"ETL démarré à {datetime.now()}")
    
    # Simulation de traitement ETL
    data = {
        'timestamp': [datetime.now()],
        'processed_records': [100],
        'status': ['success']
    }
    
    df = pd.DataFrame(data)
    print(f"Données traitées: {len(df)} enregistrements")
    
    # Ici vous pourriez sauvegarder en base de données
    # connection = psycopg2.connect(os.getenv('DATABASE_URL'))
    # ...
    
    print("ETL terminé avec succès")

if __name__ == "__main__":
    main()
