etl/pyproject.toml
[tool.poetry]
name = "dashboard-etl"
version = "0.1.0"
description = "Dashboard Analytics ETL Service"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.0"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.8.2"
pydantic = "^2.5.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
